Modified data["parllalel"]
In main(), before anything is run, if the --parallel flag has been
set data["parallel"] is set to "&" and "" otherwise. For script files,
data["parallel"] can be used in <string>.format() with no if statement
and if other kinds of parallelism is needed, this can still be used as
a conditional because (if "") == False.